Make polyglossia loadable. Closes #94.
Include hook to input a file cvloadbeforehyperref.tex before
hyperref if the file exists and lua or xelatex ist used.
This is mainly done to allow users to use and configure the
polyglossia package which needs to be loaded before  hyperref.
Explain this in the new user guide.
Of course other packages that conflict with hyperref could
be loaded in this way aswell.